poetry: update openai, llama-index-* deps
Pydantic deprecation messages originate from us using outdated library versions.

This addresses the following deprecation message on startup:
```
pydantic.warnings.UnsupportedFieldAttributeWarning: The 'validate_default' attribute with value True was provided to the `Field()` function, which has no effect in the context it was used. 'validate_default' is field-specific metadata, and can only be attached to a model field using `Annotated` metadata or by assignment. This may have happened because an `Annotated` type alias using the `type` statement was used, or if the `Field()` function was attached to a single member of a union type.
```