use pymilvus + lru + disk offloading in place of hnswlib hnswlib stores all embedding representations in memory, so it does not scale here.
chromadb internally uses hnswlib and explicitly highlights this as an issue: https://docs.trychroma.com/guides/deploy/performance
```
Chroma uses a fork of hnswlib to efficiently index and search over embedding vectors. The HNSW algorithm requires that the embedding index reside in system RAM to query or update.

As such, the amount of available system memory defines an upper bound on the size of a Chroma collection (or multiple collections, if they are being used concurrently.) If a collection grows larger than available memory, insert and query latency spike rapidly as the operating system begins swapping memory to disk. The memory layout of the index is not amenable to swapping, and the system quickly becomes unusable.

Therefore, users should always plan on having enough RAM provisioned to accommodate the anticipated total number of embeddings.
```

So we need a new third option (milvus).