feat: Add WhatsApp template messages for onboarding, leaderboard, and DYK with multi-language support (#95)
* background-jobs: deprecate POST /schedule endpoint (#93)

* background-jobs: deprecate POST /schedule endpoint

* Return 200 OK with JSON "detail" as per standards
Addresses few points made by @a4i-architect

* POST /asha_logs: tweak setup, add integration tests (#92)

* POST /asha_logs: add integration tests
test_admin.py is expected to fail at two points

* admin.py: fix POST /asha_logs not returning expected response
This change is expected to fix the failing integration tests

* test_admin: use datetime.isoformat() in place of custom date formatter

* /asha_logs: stream logs server-side, process file client-side
Unlike XLSX, CSV is one of the better formats to use here as we can iteratively deliver data to user.

XLSX is a ZIP of XML files so it cannot be streamed.

* /get_bot_messages: add a 'limit=1000' parameter
We can leverage this to get a sample of messages for later /asha_logs integration tests

* test_admin: add test case to assert /asha_logs csv output format
This test will ensure the format contains few necessary attributes and that all records are sorted by incoming timestamp attribute.
/get_bot_messages is used to fetch a sample of up to a 100 messages.

* cleanup code, diff

* workflows: temporarily run cd
This change will be reverted after failing test is fixed

* Revert "workflows: temporarily run cd"

This reverts commit dbff763d4879516d839948bf142c17351613fa09.

* changed leaderboard and send dyk code to use template instade of sending normal whatsapp message

* added template message for user to get message when he is onboarded via backed script

* final commit also checked by running background jobs and made changes to now work with template code

* Quote all DYK CSV fields to avoid comma parsing errors

* Cleanup DYK templates and enforce CSV validation

---------

Co-authored-by: Muqsit <notmuqsit+a4i@gmail.com>